[FEATURE] Add authorization to the controllers (#47)
* Inject cred and auth to api system

* Remove duplicated Authorize attributes

* Add authored to details dto

* Add ressource authorization

* Fix tests for section controller

* Add section permissions

* Add authorize to section controller

* Fix or prefix

* Fix Tietgens enormous mistakes

* Add auth to content controller

* Add authorisation checks to comment controller

* Add authroization to content controller

* Fix comment controller tests

* Change permission of section

* Remove old ressource based auth from section controller

* Remove unecessary code

* Fix formatting error

* Make test passed

* Add provider to injection

* Add provider mock in test

* Dotnet format

+++ SELearning.API/Controllers/ContentController.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Identity.Web.Resource;
+using SELearning.Core.Permission;
 
 namespace SELearning.API.Controllers;
 
@@ -12,11 +13,13 @@ public class ContentController : ControllerBase
 {
     private readonly ILogger<ContentController> _logger;
     private readonly IContentService _service;
+    private readonly IAuthorizationService _authService;
 
-    public ContentController(ILogger<ContentController> logger, IContentService service)
+    public ContentController(ILogger<ContentController> logger, IContentService service, IAuthorizationService authService)
     {
         _logger = logger;
         _service = service;
+        _authService = authService;
     }
 
     /// <summary>
@@ -24,7 +27,6 @@ public class ContentController : ControllerBase
     /// </summary>
     /// <param name="ID">The ID of the content.</param>
     /// <returns>A content with the given ID if it exists, otherwise response type 404: Not Found.</returns>
-    [Authorize]
     [HttpGet("{ID}")]
     [ProducesResponseType(typeof(ContentDto), 200)]
     [ProducesResponseType(404)]
@@ -45,7 +47,6 @@ public class ContentController : ControllerBase
     /// <c>GetAllContent</c> returns all contents.
     /// </summary>
     /// <returns>all contents if they can be found, otherwise response type 404: Not Found.</returns>
-    [Authorize]
     [HttpGet]
     [ProducesResponseType(typeof(IReadOnlyCollection<ContentDto>), 200)]
     [ProducesResponseType(404)]
@@ -61,6 +62,7 @@ public class ContentController : ControllerBase
     /// <returns>A response type 201: Created</returns>
     [HttpPost]
     [ProducesResponseType(201)]
+    [AuthorizePermission(Permission.CreateContent)]
     public async Task<IActionResult> CreateContent(ContentCreateDto content)
     {
         var createdContent = await _service.AddContent(content);
@@ -80,8 +82,17 @@ public class ContentController : ControllerBase
     {
         try
         {
-            await _service.UpdateContent(ID, content);
-            return NoContent();
+            ContentDto contentToBeUpdated = await _service.GetContent(ID);
+
+            var authResult = await _authService.AuthorizeAsync(User, contentToBeUpdated, "PermissionDeleteOwnContent OR PermissionDeleteAnyContent");
+
+            if (authResult.Succeeded)
+            {
+                await _service.UpdateContent(ID, content);
+                return NoContent();
+            }
+            else
+                return Forbid($"User is not allowed to update content with id {ID}");
         }
         catch (ContentNotFoundException)
         {
@@ -101,8 +112,17 @@ public class ContentController : ControllerBase
     {
         try
         {
-            await _service.DeleteContent(ID);
-            return NoContent();
+            ContentDto contentToBeDeleted = await _service.GetContent(ID);
+
+            var authResult = await _authService.AuthorizeAsync(User, contentToBeDeleted, "PermissionDeleteOwnContent OR PermissionDeleteAnyContent");
+
+            if (authResult.Succeeded)
+            {
+                await _service.DeleteContent(ID);
+                return NoContent();
+            }
+            else
+                return Forbid($"User is not allowed to delete content with id {ID}");
         }
         catch (ContentNotFoundException)
         {
@@ -115,10 +135,10 @@ public class ContentController : ControllerBase
     /// </summary>
     /// <param name="ID">The ID of the content.</param>
     /// <returns>A response type 204: No Content if the content exists, otherwise response type 404: Not Found.</returns>
-    [Authorize]
     [HttpPut("{ID}/Upvote")]
     [ProducesResponseType(204)]
     [ProducesResponseType(404)]
+    [AuthorizePermission(Permission.Rate)]
     public async Task<IActionResult> UpvoteContent(int ID)
     {
         try
@@ -137,10 +157,10 @@ public class ContentController : ControllerBase
     /// </summary>
     /// <param name="ID">The ID of the content.</param>
     /// <returns>A response type 204: No Content if the content exists, otherwise response type 404: Not Found.</returns>
-    [Authorize]
     [HttpPut("{ID}/Downvote")]
     [ProducesResponseType(204)]
     [ProducesResponseType(404)]
+    [AuthorizePermission(Permission.Rate)]
     public async Task<IActionResult> DownvoteContent(int ID)
     {
         try
